feat: supervisor populates fractionable_symbols at startup; order_dispatch handles float quantities
- Add get_fractionable_symbols() to FakeBroker in supervisor tests
- Call broker.get_fractionable_symbols() in RuntimeSupervisor.from_settings() and replace settings with populated frozenset
- Change int(broker_order.quantity) to float() in order_dispatch save path
- Fix %d log format specifier to %g for float quantity logging

+++ src/alpaca_bot/runtime/order_dispatch.py
@@ -339,7 +339,7 @@ def dispatch_pending_orders(
             continue
         if order.intent_type == "entry":
             logger.info(
-                "order_dispatch: entry submitted for %s — submitted_qty=%d broker_confirmed_qty=%s",
+                "order_dispatch: entry submitted for %s — submitted_qty=%g broker_confirmed_qty=%s",
                 order.symbol,
                 order.quantity,
                 broker_order.quantity,
@@ -354,7 +354,7 @@ def dispatch_pending_orders(
                         side=order.side,
                         intent_type=order.intent_type,
                         status=normalized_status,
-                        quantity=int(broker_order.quantity),
+                        quantity=float(broker_order.quantity),
                         trading_mode=order.trading_mode,
                         strategy_version=order.strategy_version,
                         strategy_name=order.strategy_name,
